+++ src/ast/visit/inventory.rs
@@ -1921,7 +1921,7 @@ macro_rules! ast_traversal {
         pub fn walk_doc_comment<V: $Visitor>(visitor: &mut V, node: &$($mutability)? Node<DocComment>) {
             visitor.enter_node(&$($mutability)? node.span);
             visitor.visit_span(&$($mutability)? node.span);
-            let DocComment { identification, locale, text } = &$($mutability)? node.value;
+            let DocComment { identification, locale, text, body_span } = &$($mutability)? node.value;
             if let Some(inner) = identification {
                 visitor.visit_identification(inner);
             }
@@ -1929,13 +1929,14 @@ macro_rules! ast_traversal {
                 visitor.visit_text(inner);
             }
             visitor.visit_text(text);
+            visitor.visit_span(body_span);
             visitor.leave_node(&$($mutability)? node.span);
         }
 
         pub fn walk_comment_annotation<V: $Visitor>(visitor: &mut V, node: &$($mutability)? Node<CommentAnnotation>) {
             visitor.enter_node(&$($mutability)? node.span);
             visitor.visit_span(&$($mutability)? node.span);
-            let CommentAnnotation { keyword_span, identification, about_targets, locale, text } = &$($mutability)? node.value;
+            let CommentAnnotation { keyword_span, identification, about_targets, locale, text, body_span } = &$($mutability)? node.value;
             if let Some(inner) = keyword_span {
                 visitor.visit_span(inner);
             }
@@ -1949,13 +1950,14 @@ macro_rules! ast_traversal {
                 visitor.visit_text(inner);
             }
             visitor.visit_text(text);
+            visitor.visit_span(body_span);
             visitor.leave_node(&$($mutability)? node.span);
         }
 
         pub fn walk_textual_representation<V: $Visitor>(visitor: &mut V, node: &$($mutability)? Node<TextualRepresentation>) {
             visitor.enter_node(&$($mutability)? node.span);
             visitor.visit_span(&$($mutability)? node.span);
-            let TextualRepresentation { rep_identification, language, language_span, text } = &$($mutability)? node.value;
+            let TextualRepresentation { rep_identification, language, language_span, text, body_span } = &$($mutability)? node.value;
             if let Some(inner) = rep_identification {
                 visitor.visit_identification(inner);
             }
@@ -1964,6 +1966,7 @@ macro_rules! ast_traversal {
                 visitor.visit_span(inner);
             }
             visitor.visit_text(text);
+            visitor.visit_span(body_span);
             visitor.leave_node(&$($mutability)? node.span);
         }